refactor : remove duplicated code among various flavours
