+++ app/src/main/res/layout-v21/single_info.xml
@@ -35,6 +35,7 @@
                 android:layout_height="wrap_content"
                 android:background="@null"
                 android:id="@+id/delete"
+                android:visibility="gone"
                 android:src="@drawable/ic_delete_black_24dp"
                 android:layout_alignParentEnd="true"
                 />
