+++ app/src/main/res/layout/single_info.xml
@@ -34,6 +34,7 @@
                 android:layout_height="wrap_content"
                 android:background="@null"
                 android:id="@+id/delete"
+                android:visibility="gone"
                 android:src="@drawable/ic_delete_black_24dp"
                 android:layout_alignParentEnd="true"
                 android:layout_alignParentRight="true" />
